full/requirements/base.txt
# Using Ubuntu 14.04 or so you should install before installing these requirements
# sudo apt-get install libjpeg-dev, python-dev
# sudo apt-get install gettext

# Remove all packages installed in PIP
# pip freeze | xargs pip uninstall -y


Django==1.8.12
Pillow==2.9.0
argparse==1.2.1
django-authtools==1.4.0
djangorestframework==3.3.3
django-rest-swagger==0.3.5
djangorestframework-gis==0.10.1
Markdown==2.6.5
django-filter==0.12.0
django-wpadmin==1.7.4
psycopg2==2.6.1
django-extensions==1.6.1
django-localflavor==1.2
django-compressor==2.0
ipdb==0.8.2
django-dbbackup==2.3.3
redis==2.10.3
django-redis-cache==1.6.4
requests==2.9.1
django-braces==1.8.1
django-guardian==1.4.1
django-filer==1.1.1
django-mptt==0.8.3
easy-thumbnails==2.3
django-polymorphic==0.8.1
raven==5.1.1
django-cors-headers==1.1.0
django-mail-templated==2.6.2